backend/requirements.txt
# Core Framework
crewai==0.80.0
crewai-tools==0.12.0

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# LLM Providers
openai==1.10.0
anthropic==0.18.0

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
